major update: in the classes.py minor syntax fixes
in input.py i moved all getters to the top of the file aswell as plan to make command choice a global varible so i can pass it into load_data.py for selection of the table to run the correct insert statements

in load_data.py i created a list and a function for creating a table with its associated arguments. I then use a sql statement class i created to build our sql statement based on the keys and  values given in our json file so it will always build the write statment and in the right order.

Problems:
 need to pass command choice from input.py file into load_data.py so we can link the choice to the table in the database.

 Need To Do:
 need to create other lists of tables for other modules. I have only done this for the account module.
 also need to create the tables themselves within the database  ****make sure naming conventions are followed*****